@@ -1,28 +1,20 @@
 <?php
+
 /**
  * @see https://github.com/dotkernel/dot-twigrenderer/ for the canonical source repository
- * @copyright Copyright (c) 2017 Apidemia (https://www.apidemia.com)
- * @license https://github.com/dotkernel/dot-twigrenderer/blob/master/LICENSE.md MIT License
  */
 
-declare(strict_types = 1);
+declare(strict_types=1);
 
 namespace Dot\Twig\Extension;
 
 use Dot\Authorization\AuthorizationInterface;
-use Psr\Http\Message\ServerRequestInterface;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
 
-/**
- * Class AuthorizationExtension
- * @package Dot\Twig\Extension
- */
 class AuthorizationExtension extends AbstractExtension
 {
-    /**
-     * @var AuthorizationInterface
-     */
+    /** @var AuthorizationInterface */
     protected $authorization;
 
     public function __construct(AuthorizationInterface $authorization)
@@ -35,9 +27,6 @@ class AuthorizationExtension extends AbstractExtension
         return 'dot-authorization';
     }
 
-    /**
-     * @return array
-     */
     public function getFunctions(): array
     {
         return [
@@ -45,11 +34,7 @@ class AuthorizationExtension extends AbstractExtension
         ];
     }
 
-    /**
-     * @param string $permission
-     * @return mixed
-     */
-    public function isGranted(string $permission = '')
+    public function isGranted(string $permission = ''): bool
     {
         return $this->authorization->isGranted($permission);
     }
